fix: return type of inw

+++ src/include/oak/io.h
@@ -3,7 +3,7 @@
 
 #include <oak/types.h>
 extern u8 inb(u16 port);
-extern u8 inw(u16 port);
+extern u16 inw(u16 port);
 
 extern void outb(u16 port, u8 value);
 extern void outw(u16 port, u16 value);
